backend/requirements.txt
# --- LLM & Agent Framework ---
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
openai>=1.30.0

# --- Vector Search ---
pinecone>=5.0.0

# --- Database ---
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# --- Data & Visualization ---
pandas>=2.2.0
tabulate>=0.9.0
matplotlib>=3.9.0
seaborn>=0.13.2  # for polished visualizations
numpy>=1.26.0

# --- Environment & Utilities ---
python-dotenv>=1.0.0
jupyterlab>=4.0.0
ipython>=8.0.0

# --- API Server ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
